backend/requirements.txt

Django>=5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3.1
psycopg2-binary>=2.9
dj-database-url>=2.2
python-dotenv>=1.0
Pillow>=10.0
cloudinary>=1.41
django-cloudinary-storage>=0.3.0
gunicorn>=21.2
whitenoise>=6.7
